+++ rotkehlchen/db/cache.py
@@ -27,6 +27,7 @@ class DBCacheStatic(Enum):
     LAST_CREATE_REMINDER_CHECK_TS: Final = 'last_create_reminder_check_ts'
     LAST_GRAPH_DELEGATIONS_CHECK_TS: Final = 'last_graph_delegations_check_ts'
     LAST_GNOSISPAY_QUERY_TS: Final = 'last_gnosispay_query_ts'
+    GNOSIS_PAY_SAFE_MIGRATION: Final = 'gnosis_pay_safe_migration'
     LAST_SPARK_ASSETS_UPDATE: Final = 'last_spark_assets_update'
     LAST_DB_UPGRADE: Final = 'last_db_upgrade'
     DOCKER_DEVICE_INFO: Final = 'docker_device_info'